requirements.txt
Pillow==9.5.0
Pillow==10.0.1
Requests==2.31.0
streamlit==1.23.1
